[AC-4764] Fix BaseUserRoleGrantEvent and Use it in Finalist Event
This commit fixed the previously created base file and then used it to convert a the UserBecameFinalistEvent.  The bones for UserBecameConfirmedMentorEvent have been lain but not completed.  To be fixed later.

The UserBecameConfirmedJudgeEvent requires additional work besides extending the base.
