+++ web/impact/impact/v1/events/__init__.py
@@ -1,3 +1,4 @@
+from impact.v1.events.base_user_role_grant_event import BaseUserRoleGrantEvent
 from impact.v1.events.organization_became_entrant_event import (
     OrganizationBecameEntrantEvent,
 )
@@ -9,8 +10,11 @@ from impact.v1.events.organization_created_event import (
     OrganizationCreatedEvent,
 )
 from impact.v1.events.user_became_confirmed_judge_event import (
-	UserBecameConfirmedJudgeEvent,
+    UserBecameConfirmedJudgeEvent,
 )
+# from impact.v1.events.user_became_confirmed_mentor_event import (
+#     UserBecameConfirmedMentorEvent,
+# )
 from impact.v1.events.user_became_finalist_event import (
     UserBecameFinalistEvent,
 )
