+++ web/impact/impact/v1/events/user_became_confirmed_mentor_event.py
@@ -1,0 +1,11 @@
+# from impact.v1.events.base_user_event import BaseUserEvent
+
+
+# class UserBecameConfirmedMentorEvent(BaseUserEvent):
+#     DESCRIPTION_FORMAT = ("Became Confirmed Judge "
+#                           "for {name} ({id})").format(
+#                                                 name=program.name,
+#                                                 id=program.id)
+#     EVENT_TYPE = "became confirmed judge"
+#     # user_role = UserRole.MENTOR
+#     result = program_role_grant.created_at
